feat(6721): split observability endpoint tests
